Simplify API
